Perfo issue while loading cityplacemark
The case is related to alphabetical sort in MarbleControlBox::m_sortproxy.
While adding the cityplacemark into the list it makes more sense
to do a reset. I didn't find the balance between simple and big case
(might be expressed as a ratio of length and size...) because the
gain is significant for now at startup.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1048253 283d02a7-25f6-0310-bc7c-ecb5cbfe19da