Add configuration option for the maximum tile level.
If maximum tile level is not configured, behaviour is like before,
that is, the autodetection code is used.
For the map themes using this option Marble does not scan the filesystem
for tile level directories repeatedly anymore and also pointless HTTP requests
for non existing tiles are prevented.

Downside of using this option is that new tile levels which might be made
available on tile servers will not be used automatically anymore.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1081980 283d02a7-25f6-0310-bc7c-ecb5cbfe19da