Fix memory management of TextureTiles.
TextureTiles are created by the TileLoader, which might need to hold a
reference to the TextureTile in case of an outdated tile until the update
arrives. The TextureTiles referenced by TileLoader (using m_waitingForUpdate)
had to be protected from destruction in ~StackedTile by checking the state of
the tile. Ownership of TextureTile was transferred from TileLoader to
StackedTile not after completion of TileLoader::loadTile, but after an
eventual update arrived.

This approach is complicated and error prone and a misuse of the tile state.
Besides it made improvements/fixes in the reload functionality quite difficult.
Using QSharePointer here fixes all these issues and make improvements for
reload possible/more easy.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1106571 283d02a7-25f6-0310-bc7c-ecb5cbfe19da