Fix (or rather improve) cost calculation for tile cache.
This is a fix for a regression introduced with the tile loader refactoring.
Since then, StackedTile is used which includes a set of tiles and the
composed result tile. This was not taken into account.

However, the cost calculation is still not correct, as in some cases
because of QImage's implicitly shared data, the calculated value is too high.
Maybe this is better than if a value too low is reported.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1106665 283d02a7-25f6-0310-bc7c-ecb5cbfe19da