dead code--: remove tile states
TextureTiles are never empty because they always hold a valid QImage (see the Q_ASSERTions in TextureTile).
Although the tile state can also be scaled or expired, the tile states are only checked for being empty.
Hence, these comparisons are needless.

Similar to the above reasoning, StackedTile::deriveCompletionState() will always result in the TileComplete state.
That way, checking for the state of a StackedTile is equivalent to checking whether it contains tiles.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1155695 283d02a7-25f6-0310-bc7c-ecb5cbfe19da