introduce MarbleWidget::viewport()
Having a convenient access to the viewport makes sense since it represents the perspective of the observer.
Realizing that MarbleMap and MarbleWidget visualize data that is present in MarbleModel, MarbleMap can be seen as an implementation detail of MarbleWidget.
That way, hiding the implementation makes sense, too.

discussed with tackat

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1165655 283d02a7-25f6-0310-bc7c-ecb5cbfe19da