Switch Marble Qt version to maps-4.5.php.
In Marble 0.10 (released with KDE 4.5) there have been changes in the
way the clouds layer is handled. It now uses the generic blending classes.
Currently there are three map themes available for download which have a clouds
layer and for which we have to provide versions for Marble versions 0.10+.

In addition to that we might (hopefully) see new map themes which use the
blending feature. These maps will not work in Marble versions before 0.10.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdeedu/marble@1167517 283d02a7-25f6-0310-bc7c-ecb5cbfe19da